Change SslContext::set_read_ahead(c_long) to SslContext::set_read_ahead(u32)

+++ openssl/src/ssl/mod.rs
@@ -423,9 +423,9 @@ impl SslContext {
         }
     }
 
-    pub fn set_read_ahead(&self, m: c_long) {
+    pub fn set_read_ahead(&self, m: u32) {
         unsafe {
-            ffi::SSL_CTX_set_read_ahead(*self.ctx, m);
+            ffi::SSL_CTX_set_read_ahead(*self.ctx, m as c_long);
         }
     }
